requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.7
django-addanother==2.0.0
django-appconf==1.0.2
django-autocomplete-light==3.3.0rc6
django-generate-secret-key==1.0.2
django-haystack==2.6.1
django-import-export==0.6.0
django-search-views==0.3.7
django-select2==6.0.0
djangorestframework==3.7.3
et-xmlfile==1.0.1
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
MarkupSafe==1.0
odfpy==1.3.5
openpyxl==2.4.9
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
